tested comm with 2 md22 drivers at same time - motor_control.py
